+++ cli/tests/pcluster/cli_commands/test_compute_fleet_status_manager.py
@@ -102,6 +102,19 @@ class TestComputeFleetStatusManager:
             (
                 ComputeFleetStatus.START_REQUESTED,
                 [
+                    ComputeFleetStatus.PROTECTED,
+                    ComputeFleetStatus.START_REQUESTED,
+                    ComputeFleetStatus.STARTING,
+                    ComputeFleetStatus.RUNNING,
+                ],
+                [None],
+                None,
+                None,
+                True,
+            ),
+            (
+                ComputeFleetStatus.START_REQUESTED,
+                [
                     ComputeFleetStatus.STOPPED,
                     ComputeFleetStatus.START_REQUESTED,
                     ComputeFleetStatus.STARTING,