+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 - Add support for Centos 7 ARM.
 - Add validation to prevent using a `cluster_resource_bucket` that is in a different region than the cluster.
 - Transition from IMDSv1 to IMDSv2.
+- Add Slurm protected mode.
 
 **CHANGES**